requirements.txt
Pillow>=9.0.0
opencv-python
numpy
scikit-image
pandas
openpyxl